+++ CoreIndustriaHuitzil/ModelsRequest/ApartadoRequest.cs
@@ -22,7 +22,7 @@ namespace CoreIndustriaHuitzil.ModelsRequest
             public string? Status { get; set; }
             public string? talla { get; set; }
             public string? articulo { get; set; }
-            public int? precio { get; set; }
+            public decimal? precio { get; set; }
             public string? type { get;  set; }
             public string? cliente { get; set;}
     }
